Initial version of HTML app builder
This is an initial version of code to combine and minify the HTML, CSS and JS.

The idea is that, when minified and combined, there will be less resources used on the ESP and lower memory usage.

This will also allow config.js to be split up in to multiple files and maybe allow conditional building to merge the code base with EmonESP, etc.

On the downside the HTML now has to be 'compiled' which can slow down development.

The minifying is done mostly in pure Python and intergrated in to the PlatformIO build process so no additional instilation steps are required.

The exception to this is the JS minification which is done using a web service (the Python minifyer is broken) so you need to be online when building.