Adds a divert_PV_ratio field which is enabled only in eco mode, with Grid import and developper mode activated. This field has a default value of 1.0 that requires 100% of PV power for charging.
Because charge cannot start with less than 6A, and is incremented by
steps of 1A, it is not possible for OpenEVSE to exactly consume the
available excess of PV power. The divert_PV_ratio declares what
minimum fraction of PV power is requireds to start charging or
increment charging current.

The default of 1.0 means that charging only starts (or is incremented)
when all the required power is available as excess PV power. A higher
value (1.1) starts with an extra safety margin.

Values below 1.0 are economically preferable if the kWh selling price
is strictly less than the kWh night-price which is the usual
situation. In this case, if the current fraction of excedent PV power
is sufficient to make the extra imported power cost less than it would
cost with the night price, it's better to charge now.

As a rule of thumb, if the Daylight price is noted "D", the Night
price "N" and the selling price "S", the optimal setting for the
divert_PV_ratio is:

		  (D-N)/(D-S)

If the actual (Daylight) price is 0.15/kWh but there is cheap Night
charging price of 0.10/kWh and an actual selling price of 0.05/kWh
then the best ratio to use is:

			(0.15 - 0.10)/(0.15 - 0.05) = 0.05/0.1 = 0.5