Add 3 sizes of larger icons for bookmarks
The icons were generated using Inkscape, with the SVG logo sourced from
https://raw.githubusercontent.com/OpenEVSE/Enclosures/master/Polycase_ML58/Enclosure%20Art.svg

Despite the name in the HTML of `apple-touch-icon`, these images are
used by a lot more than just Apple and iPhones. This should also make it
look nice as a web bookmark on Android and when a bigger icon is needed
in a desktop browser.