Remove embedded OhmConnect Vimeo video
There is really no compelling reason for this to be here, and as issue
https://github.com/OpenEVSE/ESP32_WiFi_V3.x/issues/116 points out, it
requires making multiple requests out to an external domain, unlike
anything else in the web interface.

Note that this request is made every single time you load the web
interface as well, even if you never go to the "Services" tab, since it
is all a single HTML page.